Foundational commit.
This repo is mostly meant as a place to play with tooling for a
webcomponent library that will, hopefully, stay pretty clean -- though I
worry that's already out the window...

There's already a lot going on here because these components/state
management systems where things that I was already playing with in
codepen rather than set up a whole repo, so when I finally got around to
this, a lot of it was just setting it up to work in that non-codepen
environment. That's also why you see a lot of "todos: typing" type
comments, as those issues didn't cause compilation problems in codepen.