Tabs no longer using true private fields 😢
It was breaking the build without an external dependency of @swc/helpers.

TS compiles any target <= 2021 (it's 2021 if you're from the future) to using weakmaps anyway so it was an external dependency for not a whole lot.

Once TS fixes itself I'll want to come back and revisit this library wide.