Totally refactored the state machine... again.
I decided that it was *really* annoying needing to declare getters and setters for everyhing that needed to be in state if it was just a pass through.

I also decided that I wasn't super fond of the way that I found that sometimes something was or wasn't in state object unexpectedly.

So what I've done is made it so that you need to explicitly declare the state keys when you decorate the component.

This means that we know exactly what keys we need to be watching when the constructor calls (as opposed to needing to look in the DOM and whatnot).

Because we know exactly what keys we're looking for, we can also create standard pass through getters/setters for those that don't have them declared.

I also think that this state machine is going to be easier to understand/debug, but I do need to go through and type it (and probably comment it better)

Pipes are still out.

I've also updated the current components to work with the new state machine.