Craft decorators for selecting child elements.
This is something that I had wanted to do from the get go, but hadn't been able to figure out at the time.

Saw something about the Lit element decorator that does this, which reminded me it was a thing, and convinced me to figure it out.

Also implemented the decorator in components.